refactor(desktop): split src-tauri/lib.rs into bootstrap/ by responsibility
Group the specta/invoke_handler builder pair, window geometry restoration,
the native menu, and the background task spawners into their own files
instead of one 1220-line lib.rs (#981). The specta builder pair is
`include!`d rather than `mod`-declared: `#[tauri::command]`'s hidden
`__cmd__*`/`__tauri_command_name_*` macros are `#[macro_export]`d, which
per macro_rules scoping are only auto-visible, unqualified, at the crate
root — a real module boundary there breaks every command's compilation.
